refactor: isEmpty throw의 케이스 로직 분리
