+++ BE/src/category/util/category.util.ts
@@ -1,0 +1,7 @@
+import { Category } from '../entity/category';
+import { isEmpty } from 'class-validator';
+import { CategoryNotFoundException } from '../exception/category.exception';
+
+export const validateCategory = (category: Category) => {
+  if (isEmpty(category)) throw new CategoryNotFoundException();
+};
